Refactor CBMR spatial pipeline into unified API
