+++ nimare/meta/cbmr.py
@@ -3,11 +3,13 @@
 import copy
 import logging
 import re
+import time
 from functools import wraps
 
 import numpy as np
 import pandas as pd
 import scipy
+import scipy.sparse
 
 try:
     import torch  # type: ignore[import-not-found]
@@ -32,6 +34,7 @@ from nimare.utils import (
     seed_torch,
     validate_coordinate_spaces,
 )
+from nimare.meta.utils import fit_spatial_cbmr_approximate
 
 LGR = logging.getLogger(__name__)
 __version__ = _version.get_versions()["version"]
@@ -69,7 +72,17 @@ def _normalize_named_pairwise_contrasts(contrasts):
 
 
 class CBMRResult(MetaResult):
-    """Meta-analytic result for CBMR with result-centered inference helpers."""
+    """Meta-analytic result for CBMR with result-centered inference helpers.
+
+    The same result class is used for both standard global-moderator CBMR and voxelwise
+    moderator-effect CBMR. Model-specific inference is selected from the fitted estimator's
+    ``moderator_effect`` attribute.
+    """
+
+    @property
+    def moderator_effect(self):
+        """Return the fitted moderator-effect parameterization."""
+        return getattr(self.estimator, "moderator_effect", "global")
 
     @property
     def groups(self):
@@ -80,6 +93,11 @@ class CBMRResult(MetaResult):
     def moderators(self):
         """Return fitted moderator names in display order."""
         return tuple(getattr(self.estimator, "moderators", ()) or ())
+
+    @property
+    def sv_moderator_names(self):
+        """Return spatially varying moderator map names."""
+        return tuple(name for name in self.maps if name.startswith("svModerator_"))
 
     def copy(self):
         """Return a copy of the CBMR result object."""
@@ -96,58 +114,71 @@ class CBMRResult(MetaResult):
         return new
 
     def describe_inference_inputs(self):
-        """Summarize the fitted groups and moderators for follow-up inference."""
+        """Summarize the fitted groups, moderators, and moderator-effect type."""
         return {
             "groups": self.groups,
             "moderators": self.moderators,
+            "moderator_effect": self.moderator_effect,
         }
 
-    def get_inference(self, device=None):
+    def describe_sv_effects(self):
+        """Return simple summaries for spatially varying moderator maps."""
+        return {
+            name: (float(values.min()), float(values.mean()), float(values.max()))
+            for name, values in self.maps.items()
+            if name.startswith("svModerator_")
+        }
+
+    def get_inference(self, device=None, method=None, **kwargs):
         """Return a fitted inference engine for advanced CBMR use cases."""
         inference_device = device or getattr(self.estimator, "device", "cpu")
-        inference = CBMRInference(device=inference_device)
+        inference = CBMRInference(
+            device=inference_device,
+            moderator_effect=self.moderator_effect,
+            method=method,
+            **kwargs,
+        )
         inference.fit(self)
         return inference
 
-    def infer(self, group_contrasts=False, moderator_contrasts=False, device=None):
-        """Run CBMR inference from a fitted result.
-
-        Parameters
-        ----------
-        group_contrasts : bool, dict, list, tuple, str, or None, optional
-            Group homogeneity or comparison specification. Use ``False`` to skip group inference.
-        moderator_contrasts : bool, dict, list, tuple, str, or None, optional
-            Moderator effect or comparison specification. Use ``False`` to skip moderator
-            inference.
-        device : str, optional
-            Compute device to use for inference. Defaults to the device recorded on the fitted
-            estimator.
-        """
-        inference = self.get_inference(device=device)
+    def infer(
+        self,
+        group_contrasts=False,
+        moderator_contrasts=False,
+        device=None,
+        method=None,
+        **kwargs,
+    ):
+        """Run CBMR inference from a fitted result."""
+        inference = self.get_inference(device=device, method=method, **kwargs)
         return inference.transform(
             t_con_groups=group_contrasts,
             t_con_moderators=moderator_contrasts,
         )
 
-    def test_groups(self, groups=None, device=None):
+    def test_groups(self, groups=None, device=None, method=None, **kwargs):
         """Run one-group spatial homogeneity tests for the requested groups."""
         group_contrasts = list(self.groups) if groups is None else groups
         return self.infer(
             group_contrasts=group_contrasts,
             moderator_contrasts=False,
             device=device,
+            method=method,
+            **kwargs,
         )
 
-    def compare_groups(self, contrasts, device=None):
+    def compare_groups(self, contrasts, device=None, method=None, **kwargs):
         """Run pairwise group-comparison tests using names or (group_a, group_b) tuples."""
         group_contrasts = _normalize_named_pairwise_contrasts(contrasts)
         return self.infer(
             group_contrasts=group_contrasts,
             moderator_contrasts=False,
             device=device,
+            method=method,
+            **kwargs,
         )
 
-    def test_moderators(self, moderators=None, device=None):
+    def test_moderators(self, moderators=None, device=None, method=None, **kwargs):
         """Test whether the requested moderator effects differ from zero."""
         if not self.moderators:
             raise ValueError("This CBMR result does not include experiment-level moderators.")
@@ -156,9 +187,11 @@ class CBMRResult(MetaResult):
             group_contrasts=False,
             moderator_contrasts=moderator_contrasts,
             device=device,
+            method=method,
+            **kwargs,
         )
 
-    def compare_moderators(self, contrasts, device=None):
+    def compare_moderators(self, contrasts, device=None, method=None, **kwargs):
         """Run pairwise moderator-comparison tests using names or tuples."""
         if not self.moderators:
             raise ValueError("This CBMR result does not include experiment-level moderators.")
@@ -167,6 +200,8 @@ class CBMRResult(MetaResult):
             group_contrasts=False,
             moderator_contrasts=moderator_contrasts,
             device=device,
+            method=method,
+            **kwargs,
         )
 
 
@@ -187,6 +222,11 @@ class CBMREstimator(Estimator):
     moderators : :obj:`~str` or obj:`~list` or obj:`~None`, optional
         CBMR can accommodate experiment-level moderators (e.g. sample size, year of publication).
         Default is CBMR without experiment-level moderators.
+    moderator_effect : {"voxelwise", "global"}, optional
+        How experiment-level moderator effects are parameterized. ``"global"`` fits the
+        standard CBMR model with one coefficient per moderator and group. ``"voxelwise"`` fits
+        the spatially varying CBMR backend, in which moderator effects vary smoothly over voxels.
+        Default is ``"voxelwise"``.
     model : subclass of :class:`~nimare.meta.models.GeneralLinearModelEstimator`, optional
         Stochastic model class used by CBMR. Available options are:
 
@@ -253,24 +293,91 @@ class CBMREstimator(Estimator):
 
     _required_inputs = {"coordinates": ("coordinates", None)}
     _group_column = "_cbmr_group"
+    _valid_moderator_effects = ("global", "voxelwise")
+    _valid_backends = ("full", "approximate")
+
+    @classmethod
+    def _validate_moderator_effect(cls, moderator_effect):
+        """Validate and normalize the public moderator-effect selector."""
+        if isinstance(moderator_effect, str):
+            moderator_effect = moderator_effect.lower()
+        if moderator_effect not in cls._valid_moderator_effects:
+            raise ValueError(
+                "moderator_effect must be one of "
+                f"{cls._valid_moderator_effects}. Got {moderator_effect!r}."
+            )
+        return moderator_effect
 
     def __init__(
         self,
         group_categories=None,
         moderators=None,
+        moderator_effect="voxelwise",
         mask=None,
         spline_spacing=10,
         model=models.PoissonEstimator,
         penalty=False,
+        backend="full",
         n_iter=2000,
         lr=1,
         lr_decay=0.999,
         tol=1e-9,
         device="cpu",
         random_state=None,
+        alpha=1.0,
+        damping=1e-4,
+        compute_nll=False,
         **kwargs,
     ):
+        self._init_pipeline_state(
+            group_categories=group_categories,
+            moderators=moderators,
+            moderator_effect=moderator_effect,
+            mask=mask,
+            spline_spacing=spline_spacing,
+            model=model,
+            penalty=penalty,
+            backend=backend,
+            n_iter=n_iter,
+            lr=lr,
+            lr_decay=lr_decay,
+            tol=tol,
+            device=device,
+            random_state=random_state,
+            alpha=alpha,
+            damping=damping,
+            compute_nll=compute_nll,
+            **kwargs,
+        )
+
+    def _init_pipeline_state(
+        self,
+        group_categories=None,
+        moderators=None,
+        moderator_effect="voxelwise",
+        mask=None,
+        spline_spacing=10,
+        model=models.PoissonEstimator,
+        penalty=False,
+        backend="full",
+        n_iter=2000,
+        lr=1,
+        lr_decay=0.999,
+        tol=1e-9,
+        device="cpu",
+        random_state=None,
+        alpha=1.0,
+        damping=1e-4,
+        compute_nll=False,
+        **kwargs,
+    ):
+        """Initialize shared estimator state for the selected concrete CBMR pipeline."""
         super().__init__(**kwargs)
+        self.moderator_effect = self._validate_moderator_effect(moderator_effect)
+        if backend not in self._valid_backends:
+            raise ValueError(f"backend must be one of {self._valid_backends}. Got {backend!r}.")
+        if self.moderator_effect == "voxelwise" and model is not models.PoissonEstimator:
+            raise ValueError("Voxelwise CBMR currently requires model=models.PoissonEstimator.")
         if mask is not None:
             mask = get_masker(mask)
         self.masker = mask
@@ -279,16 +386,23 @@ class CBMREstimator(Estimator):
         self.moderators = moderators
 
         self.spline_spacing = spline_spacing
-        self.model = model(
+        model_class = models.PoissonEstimator if self.moderator_effect == "voxelwise" else model
+        self.model = model_class(
             penalty=penalty, lr=lr, lr_decay=lr_decay, n_iter=n_iter, tol=tol, device=device
         )
         self.penalty = penalty
+        self.backend = backend
         self.n_iter = n_iter
         self.lr = lr
         self.lr_decay = lr_decay
         self.tol = tol
         self.device = device
         self.random_state = random_state
+        self.alpha = alpha
+        self.damping = damping
+        self.compute_nll = compute_nll
+        self.spatial_varying_model = None
+        self.spatial_varying_coef = None
         if _uses_cuda(self.device) and not torch.cuda.is_available():
             LGR.debug("cuda not found, use device cpu")
             self.device = "cpu"
@@ -315,6 +429,11 @@ class CBMREstimator(Estimator):
                             {', '.join(self.moderators)}"""
         else:
             moderators_str = ""
+        moderator_effect_str = (
+            " Moderator effects were modeled as global effects."
+            if self.moderator_effect == "global"
+            else " Moderator effects were modeled as voxelwise effects."
+        )
         if self.model.penalty:
             penalty_str = " Firth-type penalty is applied to ensure convergence."
         else:
@@ -350,7 +469,7 @@ class CBMREstimator(Estimator):
         model_description = (
             f"CBMR is a meta-regression framework that was performed with NiMARE {__version__}. "
             f"{type(self.model).__name__} model was used to model group-wise spatial intensity "
-            f"functions {moderators_str}." + model_str
+            f"functions {moderators_str}." + moderator_effect_str + model_str
         )
 
         optimization_description = (
@@ -511,12 +630,81 @@ class CBMREstimator(Estimator):
             ids_by_group,
             n_mask_voxels,
         )
-        return {
+        inputs = {
             "ids_by_group": ids_by_group,
             "moderators_by_group": moderators_by_group,
             "foci_per_voxel": foci_per_voxel,
             "foci_per_experiment": foci_per_experiment,
         }
+        if self.moderator_effect == "voxelwise":
+            inputs["foci_by_experiment_voxel"] = self._build_group_foci_matrices(
+                filtered_coordinates,
+                ids_by_group,
+                n_mask_voxels,
+            )
+        return inputs
+
+    @staticmethod
+    def _build_group_foci_matrices(coordinates, ids_by_group, n_mask_voxels):
+        """Return experiment-by-voxel foci count matrices for each group."""
+        foci_by_experiment_voxel = {}
+        if coordinates.empty:
+            for group, group_ids in ids_by_group.items():
+                foci_by_experiment_voxel[group] = scipy.sparse.csr_matrix(
+                    (len(group_ids), n_mask_voxels),
+                    dtype=np.float64,
+                )
+            return foci_by_experiment_voxel
+
+        coordinates = coordinates.loc[:, ["id", "_cbmr_mask_index"]].copy()
+        for group, group_ids in ids_by_group.items():
+            id_to_row = {exp_id: i for i, exp_id in enumerate(group_ids)}
+            group_coordinates = coordinates.loc[coordinates["id"].isin(group_ids)]
+            rows = group_coordinates["id"].map(id_to_row).to_numpy(dtype=np.int64, copy=False)
+            cols = group_coordinates["_cbmr_mask_index"].to_numpy(dtype=np.int64, copy=False)
+            data = np.ones(group_coordinates.shape[0], dtype=np.float64)
+            foci_by_experiment_voxel[group] = scipy.sparse.coo_matrix(
+                (data, (rows, cols)),
+                shape=(len(group_ids), n_mask_voxels),
+                dtype=np.float64,
+            ).tocsr()
+        return foci_by_experiment_voxel
+
+    def _as_torch_tensor(self, value):
+        """Convert an array-like object to a float64 tensor on the estimator device."""
+        if scipy.sparse.issparse(value):
+            value = value.toarray()
+        return torch.as_tensor(value, dtype=torch.float64, device=self.device)
+
+    def _prepare_torch_inputs(self):
+        """Return tensorized spatially varying CBMR inputs."""
+        bases = self._as_torch_tensor(self.inputs_["coef_spline_bases"])
+        moderators_by_group = None
+        if self.moderators:
+            moderators_by_group = {
+                group: self._as_torch_tensor(self.inputs_["moderators_by_group"][group])
+                for group in self.groups
+            }
+        foci_by_experiment_voxel = {
+            group: self._as_torch_tensor(self.inputs_["foci_by_experiment_voxel"][group])
+            for group in self.groups
+        }
+        return bases, moderators_by_group, foci_by_experiment_voxel
+
+    def _spatial_cbmr_description(self, backend):
+        """Generate a NiMARE-style description for spatially varying CBMR."""
+        if self.moderators:
+            moderator_text = (
+                "with spatially varying experiment-level moderator effects for "
+                f"{', '.join(self.moderators)}"
+            )
+        else:
+            moderator_text = "without experiment-level moderators"
+        return (
+            f"Spatially varying CBMR was performed with the {backend} backend "
+            f"{moderator_text}. The model used {len(self.groups)} group(s), "
+            f"spline spacing {self.spline_spacing}, and device {self.device}."
+        )
 
     def _build_group_foci(self, coordinates, ids_by_group, n_mask_voxels):
         """Summarize voxelwise and experiment-wise focus counts for each group."""
@@ -650,6 +838,11 @@ class CBMREstimator(Estimator):
             Support for :class:`~nimare.dataset.Dataset` inputs is deprecated and will be removed
             in a future release. Prefer :class:`~nimare.nimads.Studyset`.
         """
+        if self.moderator_effect == "voxelwise":
+            if self.backend == "approximate":
+                return self._fit_approximate(dataset)
+            return self._fit_full(dataset)
+
         init_weight_kwargs = {
             "groups": self.groups,
             "moderators": self.moderators,
@@ -671,15 +864,197 @@ class CBMREstimator(Estimator):
 
         return maps, tables, self._description_text()
 
+    def _fit_full(self, dataset):
+        """Fit spatially varying CBMR with the full torch L-BFGS backend."""
+        seed_torch(self.random_state, self.device)
+        bases, moderators_by_group, foci_by_experiment_voxel = self._prepare_torch_inputs()
+        moderators_coef_dim = len(self.moderators) if self.moderators else None
+        self.spatial_varying_model = models.SpatialCBMRModel(
+            groups=self.groups,
+            spatial_coef_dim=self.inputs_["coef_spline_bases"].shape[1],
+            moderators_coef_dim=moderators_coef_dim,
+            device=self.device,
+        )
+        optimizer = torch.optim.LBFGS(
+            params=self.spatial_varying_model.parameters(),
+            lr=self.lr,
+            max_iter=self.n_iter,
+            tolerance_change=self.tol,
+            line_search_fn="strong_wolfe",
+        )
+        scheduler = torch.optim.lr_scheduler.ExponentialLR(optimizer, gamma=self.lr_decay)
+        start_time = time.time()
+
+        def closure():
+            """Evaluate the L-BFGS closure required by torch."""
+            optimizer.zero_grad()
+            loss = self.spatial_varying_model(
+                bases,
+                moderators_by_group,
+                foci_by_experiment_voxel,
+            )
+            loss.backward()
+            return loss
+
+        optimizer.step(closure)
+        scheduler.step()
+        LGR.info("Spatially varying CBMR optimisation took %.1f s.", time.time() - start_time)
+        maps, tables = self._extract_torch_results(moderators_by_group)
+        return maps, tables, self._spatial_cbmr_description("full L-BFGS")
+
+    def _fit_approximate(self, dataset):
+        """Fit spatially varying CBMR with the approximate backend."""
+        bases = self.inputs_["coef_spline_bases"]
+        maps = {}
+        tables = {}
+        self.spatial_varying_coef = {}
+        for group in self.groups:
+            foci = self.inputs_["foci_by_experiment_voxel"][group]
+            if self.moderators:
+                moderators = self.inputs_["moderators_by_group"][group]
+            else:
+                moderators = np.empty((foci.shape[0], 0), dtype=np.float64)
+
+            augmented_moderators = np.column_stack(
+                [moderators, np.ones((foci.shape[0], 1), dtype=np.float64)]
+            )
+            coefficient = self._get_spatial_cbmr_approximate_solver()(
+                augmented_moderators,
+                bases,
+                foci,
+                tol=self.tol,
+                max_iter=self.n_iter,
+                alpha=self.alpha,
+                damping=self.damping,
+                compute_nll=self.compute_nll,
+            )
+            self.spatial_varying_coef[group] = coefficient
+            self._add_approximate_results(
+                maps,
+                tables,
+                group,
+                moderators,
+                coefficient,
+            )
+        return maps, tables, self._spatial_cbmr_description("approximate")
+
+    @staticmethod
+    def _get_spatial_cbmr_approximate_solver():
+        """Return the approximate solver used by the voxelwise backend."""
+        return fit_spatial_cbmr_approximate
+
+    def _extract_torch_results(self, moderators_by_group):
+        """Extract maps and coefficient tables from the fitted torch model."""
+        bases = self.inputs_["coef_spline_bases"]
+        maps = {}
+        tables = {}
+        for group in self.groups:
+            spatial_coef = (
+                self.spatial_varying_model.spatial_coef_linears[group]
+                .weight.detach()
+                .cpu()
+                .numpy()
+                .ravel()
+            )
+            maps[f"spatialIntensity_group-{group}"] = np.exp(bases @ spatial_coef)
+            self._add_spatial_coef_table(tables, group, spatial_coef)
+            if self.moderators:
+                moderator_coef = (
+                    self.spatial_varying_model.moderator_coef_linears[group]
+                    .weight.detach()
+                    .cpu()
+                    .numpy()
+                )
+                moderators = moderators_by_group[group].detach().cpu().numpy()
+                self._add_moderator_maps_and_tables(
+                    maps,
+                    tables,
+                    group,
+                    moderators,
+                    moderator_coef,
+                )
+        return maps, tables
+
+    @staticmethod
+    def _add_spatial_coef_table(tables, group, spatial_coef):
+        """Add one group to the CBMR-style spatial coefficient table."""
+        columns = [f"basis_{i}" for i in range(spatial_coef.size)]
+        spatial_coef_table = tables.get(
+            "spatial_regression_coef",
+            pd.DataFrame(columns=columns, dtype=np.float64),
+        )
+        spatial_coef_table = spatial_coef_table.reindex(columns=columns)
+        spatial_coef_table.loc[group] = spatial_coef
+        tables["spatial_regression_coef"] = spatial_coef_table
+
+    @staticmethod
+    def _append_group_moderator_table(tables, group, group_moderator_table):
+        """Append one group to the aggregate spatially varying moderator table."""
+        indexed_table = group_moderator_table.copy()
+        indexed_table.index = pd.MultiIndex.from_product(
+            [[group], indexed_table.index],
+            names=["group", "moderator"],
+        )
+        if "sv_moderators_regression_coef" in tables:
+            tables["sv_moderators_regression_coef"] = pd.concat(
+                [tables["sv_moderators_regression_coef"], indexed_table]
+            )
+        else:
+            tables["sv_moderators_regression_coef"] = indexed_table
+
+    @staticmethod
+    def _add_moderator_table(tables, group, moderator_names, moderator_coef):
+        """Add spatially varying moderator coefficient tables for one group."""
+        group_moderator_table = pd.DataFrame(
+            moderator_coef,
+            index=moderator_names,
+            columns=[f"basis_{i}" for i in range(moderator_coef.shape[1])],
+        )
+        tables[f"sv_moderator_regression_coef_group-{group}"] = group_moderator_table
+        CBMREstimator._append_group_moderator_table(
+            tables,
+            group,
+            group_moderator_table,
+        )
+
+    def _add_moderator_maps_and_tables(self, maps, tables, group, moderators, moderator_coef):
+        """Add spatially varying moderator maps and tables for one group."""
+        bases = self.inputs_["coef_spline_bases"]
+        for index, moderator_name in enumerate(self.moderators):
+            moderator_effect = moderators[:, index : index + 1] @ moderator_coef[index : index + 1]
+            maps[f"svModerator_{moderator_name}_group-{group}"] = (
+                moderator_effect @ bases.T
+            ).mean(axis=0)
+        maps[f"svModeratorTotal_group-{group}"] = (moderators @ moderator_coef @ bases.T).mean(
+            axis=0
+        )
+        self._add_moderator_table(tables, group, self.moderators, moderator_coef)
+
+    def _add_approximate_results(self, maps, tables, group, moderators, coefficient):
+        """Add maps and coefficient tables for one approximate-backend group."""
+        bases = self.inputs_["coef_spline_bases"]
+        n_bases = bases.shape[1]
+        coefficient = coefficient.reshape((-1, n_bases))
+        moderator_coef = coefficient[:-1]
+        spatial_coef = coefficient[-1]
+        maps[f"spatialIntensity_group-{group}"] = np.exp(bases @ spatial_coef)
+        self._add_spatial_coef_table(tables, group, spatial_coef)
+        if self.moderators:
+            self._add_moderator_maps_and_tables(
+                maps,
+                tables,
+                group,
+                moderators,
+                moderator_coef,
+            )
+
 
 class CBMRInference(object):
     """Statistical inference on fitted CBMR results.
 
     Notes
     -----
-    The CBMR API design now centers inference on :class:`CBMRResult`. This class remains the
-    lower-level implementation used by those result helpers and by advanced users who want to
-    call :meth:`fit` and :meth:`transform` directly.
+        This is the public inference entry point.
 
     .. versionadded:: 0.1.0
 
@@ -688,18 +1063,94 @@ class CBMRInference(object):
     device: :obj:`string`, optional
         Device type ('cpu' or 'cuda') represents the device on which operations will be allocated.
         Default is 'cpu'.
+    moderator_effect : {"voxelwise", "global"}, optional
+        Inference parameterization to use. ``"voxelwise"`` uses the integrated spatial CBMR
+        inference backend with sandwich or inverse-Fisher covariance estimates. ``"global"`` uses
+        the standard CBMR inference backend. Default is ``"voxelwise"``.
     """
 
-    def __init__(self, device="cpu"):
+    _valid_methods = ("sandwich", "FI")
+    _valid_sandwich_meats = ("cluster", "iid")
+    _valid_sandwich_corrections = (None, "hc0", "hc1", "hc3")
+    _voxelwise_default_method = "sandwich"
+
+    def __init__(
+        self,
+        device="cpu",
+        moderator_effect="voxelwise",
+        method=None,
+        sandwich_meat="cluster",
+        sandwich_correction="hc3",
+        ridge=1e-6,
+    ):
+        self.moderator_effect = self._normalize_moderator_effect(moderator_effect)
         self.device = device
         # device check
         if _uses_cuda(self.device) and not torch.cuda.is_available():
             LGR.debug("cuda not found, use device 'cpu'")
             self.device = "cpu"
+
         self.result = None
         self.groups = None
         self.moderators = None
+
+        if self.moderator_effect == "global":
+            self._validate_global_pipeline_options(method, sandwich_meat, sandwich_correction, ridge)
+        else: # voxelwise
+            if method is None:
+                method = "sandwich"
+            self.method = self._validate_method(method)
+            self.sandwich_meat = self._validate_sandwich_meat(sandwich_meat)
+            self.sandwich_correction = self._validate_sandwich_correction(sandwich_correction)
+            if ridge < 0:
+                raise ValueError("ridge must be nonnegative.")
+            self.ridge = ridge
+
         self._reset_inference_caches()
+
+    @classmethod
+    def _normalize_moderator_effect(cls, moderator_effect):
+        """Normalize the shared public moderator-effect selector."""
+        return CBMREstimator._validate_moderator_effect(moderator_effect)
+
+    @classmethod
+    def _validate_method(cls, method):
+        """Validate and normalize an inference standard-error method."""
+        if isinstance(method, str):
+            if method.lower() == "fi":
+                return "FI"
+            if method.lower() == "sandwich":
+                return "sandwich"
+        raise ValueError("method must be one of {'sandwich', 'FI'}.")
+
+    @classmethod
+    def _validate_sandwich_meat(cls, sandwich_meat):
+        """Validate and normalize the sandwich meat estimator."""
+        if isinstance(sandwich_meat, str):
+            sandwich_meat = sandwich_meat.lower()
+        if sandwich_meat not in cls._valid_sandwich_meats:
+            raise ValueError("sandwich_meat must be either 'cluster' or 'iid'.")
+        return sandwich_meat
+
+    @classmethod
+    def _validate_sandwich_correction(cls, sandwich_correction):
+        """Validate and normalize the sandwich leverage correction."""
+        if isinstance(sandwich_correction, str):
+            sandwich_correction = sandwich_correction.lower()
+        if sandwich_correction not in cls._valid_sandwich_corrections:
+            raise ValueError("sandwich_correction must be None, 'hc0', 'hc1', or 'hc3'.")
+        return sandwich_correction
+
+    @staticmethod
+    def _validate_global_pipeline_options(method, sandwich_meat, sandwich_correction, ridge):
+        """Reject voxelwise-only options when the global pipeline is selected."""
+        if method is not None:
+            raise ValueError("method is only supported for voxelwise moderator effects.")
+        if sandwich_meat != "cluster" or sandwich_correction != "hc3" or ridge != 1e-6:
+            raise ValueError(
+                "sandwich_meat, sandwich_correction, and ridge are only supported for "
+                "voxelwise moderator effects."
+            )
 
     def _check_fit(fn):
         """Check if CBMRInference instance has been fit."""
@@ -727,35 +1178,51 @@ class CBMRInference(object):
         return copied_result
 
     def _reset_inference_caches(self):
-        """Reset cached inference intermediates for the current fitted result."""
-        self._group_spatial_covariance_cache = {}
+        """Reset cached inference/covariance intermediates for the fitted result."""
         self._group_log_intensity_cache = {}
         self._group_null_log_intensity_cache = {}
-        self._moderator_covariance = None
-        self._moderator_variance = None
-        self._moderator_coef_table = None
+        if self.moderator_effect == "global":
+            self._group_spatial_covariance_cache = {}
+            self._moderator_covariance = None
+            self._moderator_variance = None
+            self._moderator_coef_table = None
+        else: # voxelwise
+            self._group_covariance_cache = {}
+            self._group_coefficient_cache = {}
 
     def _get_group_log_intensity(self, group):
         """Return cached group log-intensity values."""
         group_log_intensity = self._group_log_intensity_cache.get(group)
         if group_log_intensity is None:
-            group_log_intensity = np.log(self.result.maps[f"spatialIntensity_group-{group}"])
+            if self.moderator_effect == "global":
+                group_log_intensity = np.log(self.result.maps[f"spatialIntensity_group-{group}"])
+            else: # voxelwise
+                bases = self.estimator.inputs_["coef_spline_bases"]
+                coefficient = self._get_group_coefficient_matrix(group)
+                group_log_intensity = bases @ coefficient[-1]
             self._group_log_intensity_cache[group] = group_log_intensity
         return group_log_intensity
 
     def _get_group_null_log_intensity(self, group):
-        """Return cached null log-intensity for group homogeneity testing."""
+        """Return cached null baseline log-intensity for homogeneity testing."""
         group_null_log_intensity = self._group_null_log_intensity_cache.get(group)
         if group_null_log_intensity is None:
-            group_foci_per_voxel = self.estimator.inputs_["foci_per_voxel"][group]
-            group_foci_per_experiment = self.estimator.inputs_["foci_per_experiment"][group]
-            n_voxels, n_experiments = (
-                group_foci_per_voxel.shape[0],
-                group_foci_per_experiment.shape[0],
-            )
-            group_null_log_intensity = np.log(
-                np.sum(group_foci_per_voxel) / (n_voxels * n_experiments)
-            )
+            if self.moderator_effect == "global":
+                group_foci_per_voxel = self.estimator.inputs_["foci_per_voxel"][group]
+                group_foci_per_experiment = self.estimator.inputs_["foci_per_experiment"][group]
+                n_voxels, n_experiments = (
+                    group_foci_per_voxel.shape[0],
+                    group_foci_per_experiment.shape[0],
+                )
+                group_null_log_intensity = np.log(
+                    np.sum(group_foci_per_voxel) / (n_voxels * n_experiments)
+                )
+            else: # voxelwise
+                foci = self.estimator.inputs_["foci_by_experiment_voxel"][group]
+                total_foci = foci.sum()
+                n_experiments, n_voxels = foci.shape
+                group_null_log_intensity = np.log(max(float(total_foci), np.finfo(float).tiny))
+                group_null_log_intensity -= np.log(n_experiments * n_voxels)
             self._group_null_log_intensity_cache[group] = group_null_log_intensity
         return group_null_log_intensity
 
@@ -803,19 +1270,34 @@ class CBMRInference(object):
             Fitted CBMR result containing regression coefficient tables and spatial intensity
             maps.
         """
+        if self.moderator_effect == "voxelwise":
+            if not isinstance(result, CBMRResult) or result.moderator_effect != "voxelwise":
+                raise TypeError(
+                    "CBMRInference.fit with moderator_effect='voxelwise' requires a "
+                    "CBMRResult with moderator_effect='voxelwise'."
+                )
+        else:
+            if not isinstance(result, CBMRResult) or result.moderator_effect != "global":
+                raise TypeError(
+                    "CBMRInference.fit with moderator_effect='global' requires a "
+                    "CBMRResult with moderator_effect='global'."
+                )
+
         self.result = self._copy_result_for_inference(result)
         self._reset_inference_caches()
         self.estimator = self.result.estimator
-        self.estimator.device = self.device
-        self.estimator.model.device = self.device
-        self.estimator.model.to(self.device)
-        self.estimator.model._invalidate_tensor_inputs_cache()
-        self.groups = self.result.estimator.groups
-        self.moderators = self.result.estimator.moderators
-        if self.moderators:
-            self._moderator_coef_table = (
-                self.result.tables["moderators_regression_coef"].to_numpy().T
-            )
+        self.groups = list(self.result.groups)
+        self.moderators = list(self.result.moderators)
+
+        if self.moderator_effect == "global":
+            self.estimator.device = self.device
+            self.estimator.model.device = self.device
+            self.estimator.model.to(self.device)
+            self.estimator.model._invalidate_tensor_inputs_cache()
+            if self.moderators:
+                self._moderator_coef_table = (
+                    self.result.tables["moderators_regression_coef"].to_numpy().T
+                )
 
         self.create_regular_expressions()
 
@@ -922,30 +1404,25 @@ class CBMRInference(object):
         return contrast_matrix
 
     @_check_fit
-    def transform(self, t_con_groups=None, t_con_moderators=None):
-        """Conduct generalized linear hypothesis (GLH) testing on CBMR estimates.
+    def transform(self, t_con_groups=None, t_con_moderators=None, method=None):
+        """Conduct generalized linear hypothesis (GLH) testing on CBMR estimates."""
+        if self.moderator_effect == "voxelwise":
+            if method is not None:
+                validated_method = self._validate_method(method)
+                if validated_method != self.method:
+                    self.method = validated_method
+                    self._reset_inference_caches()
+            self.result.metadata["spatial_cbmr_inference_method"] = self.method
+            if self.method == "sandwich":
+                self.result.metadata["spatial_cbmr_sandwich_meat"] = self.sandwich_meat
+                self.result.metadata["spatial_cbmr_sandwich_correction"] = self.sandwich_correction
+            else:
+                self.result.metadata.pop("spatial_cbmr_sandwich_meat", None)
+                self.result.metadata.pop("spatial_cbmr_sandwich_correction", None)
+        else:
+            if method is not None:
+                raise ValueError("method is only supported for voxelwise moderator effects.")
 
-        Estimate group-wise spatial regression coefficients and its standard
-        error via inverse Fisher Information matrix, estimate standard error of
-        group-wise log intensity, group-wise intensity via delta method. For NB
-        or clustered model, estimate regression coefficient of overdispersion.
-        Similarly, estimate regression coefficient of experiment-level
-        moderators (if exist), as well as its standard error via Fisher
-        Information matrix. Save these outcomes in `tables`. Also, estimate
-        group-wise spatial intensity (per experiment) and save the results in
-        `maps`.
-
-        Parameters
-        ----------
-        t_con_groups : bool, dict, list, tuple, str, or None, optional
-            Group inference specification. Use ``None`` or ``True`` to test all groups,
-            ``False`` to skip group inference, named contrasts such as ``"group_a-group_b"`` or
-            ``("group_a", "group_b")`` for pairwise tests, a dict mapping names to contrast
-            arrays, or raw contrast arrays.
-        t_con_moderators : bool, dict, list, tuple, str, or None, optional
-            Moderator inference specification with the same accepted forms as
-            ``t_con_groups``.
-        """
         self.t_con_groups = t_con_groups
         self.t_con_moderators = t_con_moderators
 
@@ -960,10 +1437,10 @@ class CBMRInference(object):
 
         return self.result
 
-    def fit_transform(self, result, t_con_groups=None, t_con_moderators=None):
+    def fit_transform(self, result, t_con_groups=None, t_con_moderators=None, method=None):
         """Fit and transform."""
         self.fit(result)
-        return self.transform(t_con_groups, t_con_moderators)
+        return self.transform(t_con_groups, t_con_moderators, method=method)
 
     @_check_fit
     def _preprocess_t_con_regressor(self, source):
@@ -1129,7 +1606,12 @@ class CBMRInference(object):
             involved_groups,
             is_homogeneity_test,
         )
-        cov_spatial_coef = self._get_group_spatial_covariance(involved_groups)
+        if self.moderator_effect == "voxelwise":
+            cov_spatial_coef = self._get_intercept_covariance_for_groups(involved_groups)
+            spatial_coef_dim = None
+            n_brain_voxel = None
+        else:
+            cov_spatial_coef = self._get_group_spatial_covariance(involved_groups)
 
         if con_group.shape[0] == 1:
             z_stats_spatial, p_vals_spatial = self._compute_group_wald_statistics(
@@ -1148,9 +1630,9 @@ class CBMRInference(object):
                 cov_spatial_coef,
                 contrast_log_intensity,
                 X,
-                spatial_coef_dim,
-                n_brain_voxel,
-                is_homogeneity_test,
+                spatial_coef_dim=spatial_coef_dim,
+                n_brain_voxel=n_brain_voxel,
+                is_homogeneity_test=is_homogeneity_test,
             )
 
         return {
@@ -1195,9 +1677,11 @@ class CBMRInference(object):
         cov_spatial_coef,
         contrast_log_intensity,
         X,
-        spatial_coef_dim,
+        spatial_coef_dim=None,
     ):
         """Compute one-contrast Wald statistics for group inference."""
+        if spatial_coef_dim is None:
+            spatial_coef_dim = X.shape[1]
         n_con_group_involved = len(involved_groups)
         var_log_intensity = []
         for k in range(n_con_group_involved):
@@ -1209,8 +1693,12 @@ class CBMRInference(object):
             var_log_intensity.append(var_log_intensity_k)
         var_log_intensity = np.stack(var_log_intensity, axis=0)
         involved_var_log_intensity = simp_con_group**2 @ var_log_intensity
-        involved_std_log_intensity = np.sqrt(involved_var_log_intensity)
-        z_stats_spatial = contrast_log_intensity / involved_std_log_intensity
+        involved_std_log_intensity = np.sqrt(np.maximum(involved_var_log_intensity, 0.0))
+        z_stats_spatial = contrast_log_intensity / np.where(
+            involved_std_log_intensity > 0,
+            involved_std_log_intensity,
+            np.inf,
+        )
         if n_con_group_involved == 1:
             p_vals_spatial = scipy.stats.norm.sf(z_stats_spatial)
         else:
@@ -1229,13 +1717,17 @@ class CBMRInference(object):
         cov_spatial_coef,
         contrast_log_intensity,
         X,
-        spatial_coef_dim,
-        n_brain_voxel,
-        is_homogeneity_test,
+        spatial_coef_dim=None,
+        n_brain_voxel=None,
+        is_homogeneity_test=False,
     ):
         """Compute multi-row GLH statistics for group inference."""
         n_con_group_involved = len(involved_groups)
         m = simp_con_group.shape[0]
+        if spatial_coef_dim is None:
+            spatial_coef_dim = X.shape[1]
+        if n_brain_voxel is None:
+            n_brain_voxel = X.shape[0]
         cov_spatial_coef = cov_spatial_coef.reshape(
             n_con_group_involved,
             spatial_coef_dim,
@@ -1303,14 +1795,16 @@ class CBMRInference(object):
         """
         self._run_group_inference()
 
+    @staticmethod
     def _chi_square_log_intensity(
-        self,
-        m,
-        n_brain_voxel,
-        n_con_group_involved,
-        simp_con_group,
-        cov_log_intensity,
-        contrast_log_intensity,
+        m=None,
+        n_brain_voxel=None,
+        n_con_group_involved=None,
+        simp_con_group=None,
+        cov_log_intensity=None,
+        contrast_log_intensity=None,
+        n_voxels=None,
+        n_involved_groups=None,
     ):
         """
         Calculate chi-square statistics for GLH on group-wise log intensity function.
@@ -1338,6 +1832,10 @@ class CBMRInference(object):
             Voxel-wise chi-square statistics for GLH tests on group-wise spatial
             intensity estimations.
         """
+        if n_voxels is not None:
+            n_brain_voxel = n_voxels
+        if n_involved_groups is not None:
+            n_con_group_involved = n_involved_groups
         if cov_log_intensity.ndim == 3:
             if cov_log_intensity.shape[:2] == (n_con_group_involved, n_con_group_involved):
                 cov_by_voxel = np.moveaxis(cov_log_intensity, -1, 0)
@@ -1361,6 +1859,17 @@ class CBMRInference(object):
     @_check_fit
     def _run_moderator_inference(self):
         """Evaluate all prepared moderator contrasts and write them into the result object."""
+        if self.moderator_effect == "voxelwise":
+            for con_moderator_count, con_moderator in enumerate(self.t_con_moderators):
+                for group in self.groups:
+                    moderator_stats = self._evaluate_moderator_contrast(group, con_moderator)
+                    self._store_moderator_inference_result(
+                        con_moderator_count,
+                        group,
+                        moderator_stats,
+                    )
+            return
+
         cov_moderator_coef, var_moderator_coef = self._get_moderator_covariance()
         moderator_coef = self._moderator_coef_table
         for con_moderator_count, con_moderator in enumerate(self.t_con_moderators):
@@ -1372,14 +1881,29 @@ class CBMRInference(object):
             )
             self._store_moderator_inference_result(con_moderator_count, moderator_stats)
 
-    @staticmethod
     def _evaluate_moderator_contrast(
+        self,
         con_moderator,
-        cov_moderator_coef,
-        var_moderator_coef,
-        moderator_coef,
+        cov_moderator_coef=None,
+        var_moderator_coef=None,
+        moderator_coef=None,
     ):
         """Compute statistics for one prepared moderator contrast."""
+        if self.moderator_effect == "voxelwise":
+            group = con_moderator
+            con_moderator = cov_moderator_coef
+            bases = self.estimator.inputs_["coef_spline_bases"]
+            coefficient = self._get_group_coefficient_matrix(group)
+            covariance = self._get_group_covariance(group)
+            intercept_column = np.zeros((con_moderator.shape[0], 1))
+            augmented_contrast = np.column_stack([con_moderator, intercept_column])
+            return self._compute_spatial_coefficient_statistics(
+                coefficient,
+                covariance,
+                augmented_contrast,
+                bases,
+            )
+
         m_con_moderator = con_moderator.shape[0]
         contrast_moderator_coef = np.matmul(con_moderator, moderator_coef)
         if m_con_moderator == 1:
@@ -1416,8 +1940,45 @@ class CBMRInference(object):
             "z": z_stats_moderator,
         }
 
-    def _store_moderator_inference_result(self, con_moderator_count, moderator_stats):
+    def _store_moderator_inference_result(
+        self,
+        con_moderator_count,
+        moderator_stats_or_group,
+        moderator_stats=None,
+    ):
         """Write one computed moderator-inference result into result tables."""
+        if moderator_stats is not None:
+            group = moderator_stats_or_group
+            contrast_name = (
+                self.t_con_moderators_name[con_moderator_count]
+                if self.t_con_moderators_name
+                else None
+            )
+            if contrast_name:
+                if moderator_stats["contrast_count"] > 1:
+                    self.result.maps[f"chiSquare_svModerator_{contrast_name}_group-{group}"] = (
+                        moderator_stats["chi_square"]
+                    )
+                self.result.maps[f"p_svModerator_{contrast_name}_group-{group}"] = (
+                    moderator_stats["p"]
+                )
+                self.result.maps[f"z_svModerator_{contrast_name}_group-{group}"] = (
+                    moderator_stats["z"]
+                )
+            else:
+                if moderator_stats["contrast_count"] > 1:
+                    self.result.maps[
+                        f"chiSquare_GLH_svModerators_{con_moderator_count}_group-{group}"
+                    ] = moderator_stats["chi_square"]
+                self.result.maps[f"p_GLH_svModerators_{con_moderator_count}_group-{group}"] = (
+                    moderator_stats["p"]
+                )
+                self.result.maps[f"z_GLH_svModerators_{con_moderator_count}_group-{group}"] = (
+                    moderator_stats["z"]
+                )
+            return
+
+        moderator_stats = moderator_stats_or_group
         contrast_name = (
             self.t_con_moderators_name[con_moderator_count] if self.t_con_moderators_name else None
         )
@@ -1462,3 +2023,308 @@ class CBMRInference(object):
         effects across multiple moderator effects.
         """
         self._run_moderator_inference()
+
+    def _get_group_augmented_moderators(self, group):
+        """Return the group moderator matrix with an intercept as the final column."""
+        if self.moderators:
+            moderators = np.asarray(self.estimator.inputs_["moderators_by_group"][group])
+        else:
+            n_experiments = self.estimator.inputs_["foci_by_experiment_voxel"][group].shape[0]
+            moderators = np.empty((n_experiments, 0), dtype=np.float64)
+        return np.column_stack([moderators, np.ones((moderators.shape[0], 1))])
+
+    def _get_group_coefficient_matrix(self, group):
+        """Return a ``(n_moderators + 1, n_bases)`` coefficient matrix."""
+        coefficient = self._group_coefficient_cache.get(group)
+        if coefficient is not None:
+            return coefficient
+
+        n_bases = self.estimator.inputs_["coef_spline_bases"].shape[1]
+        if getattr(self.estimator, "spatial_varying_coef", None) is not None:
+            coefficient = np.asarray(self.estimator.spatial_varying_coef[group]).reshape(
+                (-1, n_bases)
+            )
+        elif getattr(self.estimator, "spatial_varying_model", None) is not None:
+            spatial_coef = (
+                self.estimator.spatial_varying_model.spatial_coef_linears[group]
+                .weight.detach()
+                .cpu()
+                .numpy()
+                .ravel()
+            )
+            if self.moderators:
+                moderator_coef = (
+                    self.estimator.spatial_varying_model.moderator_coef_linears[group]
+                    .weight.detach()
+                    .cpu()
+                    .numpy()
+                )
+                coefficient = np.vstack([moderator_coef, spatial_coef])
+            else:
+                coefficient = spatial_coef.reshape((1, -1))
+        else:
+            spatial_coef = self.result.tables["spatial_regression_coef"].loc[group].to_numpy()
+            if self.moderators:
+                moderator_coef = self.result.tables[
+                    f"sv_moderator_regression_coef_group-{group}"
+                ].to_numpy()
+                coefficient = np.vstack([moderator_coef, spatial_coef])
+            else:
+                coefficient = spatial_coef.reshape((1, -1))
+
+        self._group_coefficient_cache[group] = coefficient
+        return coefficient
+
+    def _get_group_mean(self, group):
+        """Return fitted Poisson mean for one group as experiments by voxels."""
+        bases = self.estimator.inputs_["coef_spline_bases"]
+        moderators = self._get_group_augmented_moderators(group)
+        coefficient = self._get_group_coefficient_matrix(group)
+        return np.exp(np.clip(moderators @ coefficient @ bases.T, -100, 100))
+
+    @staticmethod
+    def _compute_fisher_information(moderators, bases, mean):
+        """Compute Fisher information for the Kronecker design without forming it."""
+        n_moderators = moderators.shape[1]
+        n_bases = bases.shape[1]
+        fisher_info = np.zeros((n_moderators * n_bases, n_moderators * n_bases))
+        moderator_weights = np.einsum("mr,ms,mv->rsv", moderators, moderators, mean)
+        for row in range(n_moderators):
+            row_slice = slice(row * n_bases, (row + 1) * n_bases)
+            for col in range(row, n_moderators):
+                col_slice = slice(col * n_bases, (col + 1) * n_bases)
+                block = bases.T @ (bases * moderator_weights[row, col, :, None])
+                fisher_info[row_slice, col_slice] = block
+                if row != col:
+                    fisher_info[col_slice, row_slice] = block.T
+        return fisher_info
+
+    @staticmethod
+    def _as_dense_response(foci):
+        """Return a dense experiment-by-voxel response matrix."""
+        if scipy.sparse.issparse(foci):
+            return foci.toarray()
+        return np.asarray(foci, dtype=float)
+
+    @staticmethod
+    def _sandwich_bread_inverse(fisher_info, ridge):
+        """Invert the sandwich bread term with a ridge-stabilized pseudo-inverse."""
+        regularized = fisher_info + ridge * np.eye(fisher_info.shape[0])
+        bread_inverse = np.linalg.pinv(regularized)
+        return 0.5 * (bread_inverse + bread_inverse.T)
+
+    @staticmethod
+    def _sandwich_meat_matrix(moderators, bases, residuals, meat):
+        """Compute cluster- or iid-robust sandwich meat for the Kronecker design."""
+        n_experiments, n_moderators = moderators.shape
+        n_bases = bases.shape[1]
+        n_parameters = n_moderators * n_bases
+        if meat == "cluster":
+            basis_residuals = bases.T @ residuals.T
+            cluster_scores = np.zeros((n_parameters, n_experiments), dtype=float)
+            for moderator_index in range(n_moderators):
+                parameter_slice = slice(
+                    moderator_index * n_bases,
+                    (moderator_index + 1) * n_bases,
+                )
+                cluster_scores[parameter_slice, :] = (
+                    basis_residuals * moderators[:, moderator_index][None, :]
+                )
+            return cluster_scores @ cluster_scores.T
+
+        residual_weights = np.einsum("mr,ms,mv->rsv", moderators, moderators, residuals**2)
+        meat_matrix = np.zeros((n_parameters, n_parameters), dtype=float)
+        for row in range(n_moderators):
+            row_slice = slice(row * n_bases, (row + 1) * n_bases)
+            for col in range(row, n_moderators):
+                col_slice = slice(col * n_bases, (col + 1) * n_bases)
+                block = bases.T @ (bases * residual_weights[row, col, :, None])
+                meat_matrix[row_slice, col_slice] = block
+                if row != col:
+                    meat_matrix[col_slice, row_slice] = block.T
+        return meat_matrix
+
+    @classmethod
+    def _apply_sandwich_correction(
+        cls,
+        correction,
+        bread_inverse,
+        moderators,
+        bases,
+        mean,
+        residuals,
+    ):
+        """Apply HC-style finite-sample corrections to sandwich residuals."""
+        if correction is None or correction == "hc0":
+            return residuals, 1.0
+
+        n_experiments, n_moderators = moderators.shape
+        if correction == "hc1":
+            if n_experiments <= n_moderators:
+                raise ValueError(
+                    "HC1 sandwich correction requires more experiments than model columns. "
+                    "Use sandwich_correction='hc0' or 'hc3' for this setting."
+                )
+            return residuals, n_experiments / float(n_experiments - n_moderators)
+
+        n_bases = bases.shape[1]
+        bread_inverse_blocks = bread_inverse.reshape(
+            n_moderators,
+            n_bases,
+            n_moderators,
+            n_bases,
+        ).transpose(0, 2, 1, 3)
+        leverage_basis = np.einsum(
+            "vp,rspq,vq->rsv",
+            bases,
+            bread_inverse_blocks,
+            bases,
+            optimize=True,
+        )
+        leverage = mean * np.einsum(
+            "mr,ms,rsv->mv",
+            moderators,
+            moderators,
+            leverage_basis,
+            optimize=True,
+        )
+        leverage = np.nan_to_num(leverage, nan=0.0, posinf=1.0, neginf=0.0)
+        leverage = np.clip(leverage, 0.0, 0.999)
+        return residuals / np.maximum(1.0 - leverage, 1e-6), 1.0
+
+    @classmethod
+    def _compute_sandwich_covariance(
+        cls,
+        moderators,
+        bases,
+        foci,
+        mean,
+        ridge=1e-6,
+        meat="cluster",
+        correction="hc3",
+    ):
+        """Compute robust Poisson sandwich covariance for one spatial CBMR group."""
+        moderators = np.asarray(moderators, dtype=float)
+        bases = np.asarray(bases, dtype=float)
+        y = cls._as_dense_response(foci)
+        mean = np.asarray(mean, dtype=float)
+        mean = np.nan_to_num(mean, nan=0.0, posinf=1e12, neginf=0.0)
+        mean = np.clip(mean, 1e-12, 1e12)
+
+        if y.shape != mean.shape:
+            raise ValueError("foci and mean must have matching experiment-by-voxel shapes.")
+        if y.shape != (moderators.shape[0], bases.shape[0]):
+            raise ValueError("foci must have shape (n_experiments, n_voxels).")
+
+        fisher_info = cls._compute_fisher_information(moderators, bases, mean)
+        bread_inverse = cls._sandwich_bread_inverse(fisher_info, ridge)
+        residuals = np.nan_to_num(y - mean, nan=0.0, posinf=0.0, neginf=0.0)
+        residuals, correction_factor = cls._apply_sandwich_correction(
+            correction,
+            bread_inverse,
+            moderators,
+            bases,
+            mean,
+            residuals,
+        )
+        meat_matrix = cls._sandwich_meat_matrix(moderators, bases, residuals, meat)
+        covariance = correction_factor * bread_inverse @ meat_matrix @ bread_inverse
+        return 0.5 * (covariance + covariance.T)
+
+    def _get_group_covariance(self, group):
+        """Return cached covariance of augmented coefficients for one group."""
+        cache_key = (group, self.method, self.sandwich_meat, self.sandwich_correction, self.ridge)
+        covariance = self._group_covariance_cache.get(cache_key)
+        if covariance is not None:
+            return covariance
+
+        bases = self.estimator.inputs_["coef_spline_bases"]
+        moderators = self._get_group_augmented_moderators(group)
+        mean = self._get_group_mean(group)
+        if self.method == "FI":
+            fisher_info = self._compute_fisher_information(moderators, bases, mean)
+            covariance = self._sandwich_bread_inverse(fisher_info, self.ridge)
+        else:
+            covariance = self._compute_sandwich_covariance(
+                moderators,
+                bases,
+                self.estimator.inputs_["foci_by_experiment_voxel"][group],
+                mean,
+                ridge=self.ridge,
+                meat=self.sandwich_meat,
+                correction=self.sandwich_correction,
+            )
+        self._group_covariance_cache[cache_key] = covariance
+        return covariance
+
+    def _get_intercept_covariance_for_groups(self, involved_groups):
+        """Return block-diagonal covariance for group spatial-intercept coefficients."""
+        n_bases = self.estimator.inputs_["coef_spline_bases"].shape[1]
+        covariance = np.zeros((len(involved_groups) * n_bases, len(involved_groups) * n_bases))
+        for group_index, group in enumerate(involved_groups):
+            group_covariance = self._get_group_covariance(group)
+            block = group_covariance[-n_bases:, -n_bases:]
+            group_slice = slice(group_index * n_bases, (group_index + 1) * n_bases)
+            covariance[group_slice, group_slice] = block
+        return covariance
+
+    @staticmethod
+    def _contrast_covariance_by_voxel(contrast, covariance, bases):
+        """Project coefficient covariance into voxel-wise contrast covariance."""
+        n_regressors = contrast.shape[1]
+        n_bases = bases.shape[1]
+        covariance_blocks = covariance.reshape(n_regressors, n_bases, n_regressors, n_bases)
+        contrast_covariance_blocks = np.einsum(
+            "sr,rpuq,tu->sptq",
+            contrast,
+            covariance_blocks,
+            contrast,
+            optimize=True,
+        )
+        return np.einsum(
+            "np,sptq,nq->nst",
+            bases,
+            contrast_covariance_blocks,
+            bases,
+            optimize=True,
+        )
+
+    @classmethod
+    def _compute_spatial_coefficient_statistics(cls, coefficient, covariance, contrast, bases):
+        """Compute voxel-wise Wald statistics for spatially varying coefficients."""
+        contrast_eta = contrast @ coefficient @ bases.T
+        contrast_cov = cls._contrast_covariance_by_voxel(contrast, covariance, bases)
+
+        if contrast.shape[0] == 1:
+            contrast_var = contrast_cov[:, 0, 0]
+            contrast_std = np.sqrt(np.maximum(contrast_var, 0.0))
+            z_stats = contrast_eta[0] / np.where(contrast_std > 0, contrast_std, np.inf)
+            p_vals = scipy.stats.norm.sf(np.abs(z_stats)) * 2
+            p_vals = _clip_p_values(
+                p_vals,
+                dtype=np.asarray(p_vals).dtype,
+                copy=False,
+            )
+            chi_square = None
+        else:
+            solved = np.linalg.solve(contrast_cov, contrast_eta.T[..., np.newaxis])
+            chi_square = np.einsum("ns,ns->n", contrast_eta.T, solved[..., 0], optimize=True)
+            p_vals = scipy.stats.chi2.sf(chi_square, df=contrast.shape[0])
+            p_vals = _clip_p_values(
+                p_vals,
+                dtype=np.asarray(p_vals).dtype,
+                copy=False,
+            )
+            z_p_values = np.maximum(
+                p_vals,
+                2 * _minimum_positive_float(np.asarray(p_vals).dtype),
+            )
+            z_stats = scipy.stats.norm.isf(z_p_values / 2)
+
+        return {
+            "contrast_count": contrast.shape[0],
+            "chi_square": chi_square,
+            "p": p_vals,
+            "z": z_stats,
+        }
+
